+++ src/main/java/nextstep/subway/ui/PathController.java
@@ -1,8 +1,8 @@
 package nextstep.subway.ui;
 
 import nextstep.subway.applicaion.PathService;
-import nextstep.subway.domain.PathType;
 import nextstep.subway.applicaion.dto.PathResponse;
+import nextstep.subway.domain.PathType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -17,7 +17,7 @@ public class PathController {
     }
 
     @GetMapping("/paths")
-    public ResponseEntity<PathResponse> findPath(@RequestParam Long source, @RequestParam Long target, @RequestParam("type") PathType type) {
+    public ResponseEntity<PathResponse> findPath(@RequestParam Long source, @RequestParam Long target, @RequestParam PathType type) {
         return ResponseEntity.ok(pathService.findPath(source, target, type));
     }
 }
