+++ src/test/java/nextstep/subway/acceptance/AcceptanceTest.java
@@ -9,7 +9,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.web.server.LocalServerPort;
 import org.springframework.test.context.ActiveProfiles;
 
-import static nextstep.subway.acceptance.MemberSteps.로그인_되어_있음;
+import static nextstep.subway.acceptance.member.MemberSteps.로그인_되어_있음;
 
 @ActiveProfiles("test")
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
@@ -26,7 +26,7 @@ public class AcceptanceTest {
     @Autowired
     private DataLoader dataLoader;
 
-    String 관리자;
+    protected String 관리자;
 
     @BeforeEach
     public void setUp() {
