+++ src/test/java/nextstep/subway/acceptance/favorite/FavoriteSteps.java
@@ -1,4 +1,4 @@
-package nextstep.subway.acceptance;
+package nextstep.subway.acceptance.favorite;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
