+++ src/test/java/nextstep/subway/acceptance/member/MemberSteps.java
@@ -1,8 +1,9 @@
-package nextstep.subway.acceptance;
+package nextstep.subway.acceptance.member;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
+import nextstep.subway.acceptance.AcceptanceTestSteps;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
